feat: Introduce QuickAccess control and update UI layout
Updated MainWindow backdrop and restructured HomePage layout.
Removed old button handlers and added QuickAccess control for favorites, shared items, and trash.
Updated resource keys and added new SVG icons.
