+++ src/gui4/windows/kDrive client/kDrive client/Pages/HomePage.xaml
@@ -29,104 +29,50 @@
     <Page.Resources>
         <converters:BooleanToInvertedVisibilityConverter  x:Key="BooleanToInvertedVisibilityConverter" />
     </Page.Resources>
-    <Grid Margin="24"
-          RowSpacing="24"
-          ColumnSpacing="24">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="2.3*" />
-            <ColumnDefinition Width="1*" />
-        </Grid.ColumnDefinitions>
+    <Grid Margin="{StaticResource Infomaniak.Style.Thickness.Uniform.L}">
         <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
             <!-- Update banner -->
             <RowDefinition Height="Auto" />
-            <!-- Sync status bar -->
+            <!-- Main -->
             <RowDefinition Height="*" />
-            <!-- Main content -->
-            <RowDefinition Height="Auto" />
             <!-- Upgrade offer bar -->
+            <RowDefinition Height="Auto" />
         </Grid.RowDefinitions>
 
         <!-- Update Available Banner -->
         <controls:UpdateAvailableBar Grid.Row="0"
-                                     Grid.ColumnSpan="2"
-                                     Visibility="Visible" />
+                                     Visibility="Collapsed"
+                                     Margin="{StaticResource Infomaniak.Style.Thickness.Bottom.M}" />
 
-        <!-- Sync Status Bar -->
-        <controls:SyncStatusBar Grid.Row="1" />
-
-        <!-- Left main pane -->
-        <controls:SyncActivityTable Grid.Row="2"
-                                    Grid.Column="0" />
-
-        <!-- Right Side Buttons -->
-        <Grid Grid.Row="2"
-              Grid.Column="1">
+        <!-- Main content-->
+        <Grid Grid.Row="1"
+              RowSpacing="{StaticResource Infomaniak.Style.Spacing.L}"
+              ColumnSpacing="{StaticResource Infomaniak.Style.Spacing.M}">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="2.3*" />
+                <ColumnDefinition Width="1*" />
+            </Grid.ColumnDefinitions>
             <Grid.RowDefinitions>
-                <RowDefinition Height="*" />
-                <RowDefinition Height="*" />
+                <!-- Sync status bar -->
+                <RowDefinition Height="Auto" />
+                <!-- Upgrade offer bar -->
                 <RowDefinition Height="*" />
             </Grid.RowDefinitions>
-            <Button Click="FavoritesButton_Click"
-                    CornerRadius="8"
-                    Padding="0"
-                    Grid.Row="0"
-                    VerticalAlignment="Stretch"
-                    HorizontalAlignment="Stretch"
-                    Margin="0,0,0,8"
-                    Background="{ThemeResource LayerFillColorDefaultBrush}">
-                <StackPanel Orientation="Vertical">
-                    <ImageIcon Height="16"
-                               Margin="0,0,0,8">
-                        <ImageIcon.Source>
-                            <SvgImageSource UriSource="ms-appx:///Assets/Icons/star.svg" />
-                        </ImageIcon.Source>
-                    </ImageIcon>
-                    <TextBlock x:Uid="Page_HomePage_favorites"
-                               Style="{StaticResource Infomaniak.Style.TextBlock.Body}" />
-                </StackPanel>
-            </Button>
-            <Button Click="SharedButton_Click"
-                    CornerRadius="8"
-                    Padding="0"
-                    Grid.Row="1"
-                    VerticalAlignment="Stretch"
-                    HorizontalAlignment="Stretch"
-                    Margin="0,8,0,8"
-                    Background="{ThemeResource LayerFillColorDefaultBrush}">
-                <StackPanel Orientation="Vertical">
-                    <ImageIcon Height="16"
-                               Margin="0,0,0,8">
-                        <ImageIcon.Source>
-                            <SvgImageSource UriSource="ms-appx:///Assets/Icons/share.svg" />
-                        </ImageIcon.Source>
-                    </ImageIcon>
-                    <TextBlock x:Uid="Page_HomePage_shared"
-                               Style="{StaticResource Infomaniak.Style.TextBlock.Body}" />
-                </StackPanel>
-            </Button>
-            <Button Click="TrashButton_Click"
-                    CornerRadius="8"
-                    Padding="0"
-                    Grid.Row="2"
-                    VerticalAlignment="Stretch"
-                    HorizontalAlignment="Stretch"
-                    Margin="0,8,0,0"
-                    Background="{ThemeResource LayerFillColorDefaultBrush}">
-                <StackPanel Orientation="Vertical">
-                    <ImageIcon Height="16"
-                               Margin="0,0,0,8">
-                        <ImageIcon.Source>
-                            <SvgImageSource UriSource="ms-appx:///Assets/Icons/trash.svg" />
-                        </ImageIcon.Source>
-                    </ImageIcon>
-                    <TextBlock x:Uid="Page_HomePage_trash"
-                               Style="{StaticResource Infomaniak.Style.TextBlock.Body}" />
-                </StackPanel>
-            </Button>
+
+            <!-- Sync Status Bar -->
+            <controls:SyncStatusBar Grid.Row="0" />
+
+            <!-- Left main pane -->
+            <controls:SyncActivityTable Grid.Row="1"
+                                        Grid.Column="0" />
+
+            <!-- Right Side Buttons -->
+            <controls:QuickAccess Grid.Row="1"
+                                  Grid.Column="1" />
         </Grid>
-        <controls:KSuiteUpgradeBar Grid.Row="3"
+        <controls:KSuiteUpgradeBar Grid.Row="2"
                                    Grid.ColumnSpan="2"
-                                   Visibility="{x:Bind ViewModel.SelectedSync.Drive.IsPaidOffer, Mode=OneWay, Converter={StaticResource BooleanToInvertedVisibilityConverter }}" />
+                                   Visibility="{x:Bind ViewModel.SelectedSync.Drive.IsPaidOffer, Mode=OneWay, Converter={StaticResource BooleanToInvertedVisibilityConverter }}"
+                                   Margin="{StaticResource Infomaniak.Style.Thickness.Top.M}" />
     </Grid>
 </Page>
