+++ src/gui4/windows/kDrive client/kDrive client/kDrive client.csproj
@@ -69,6 +69,9 @@
     <AppxPackageSigningEnabled>True</AppxPackageSigningEnabled>
   </PropertyGroup>
   <ItemGroup>
+    <Page Update="CustomControls\QuickAccess.xaml">
+      <SubType>Designer</SubType>
+    </Page>
     <Page Update="CustomControls\DriveIcon.xaml">
       <SubType>Designer</SubType>
     </Page>
